W471: mnemonic-signature census; the defects are in the declarations
Ten lanes against a new instrument that classifies every NONMATCHING function
by its dominant semantic instruction-selection delta.

Result: 6 functions improved, 1 regressed by 2 bytes in the same file as a
larger gain, nd -245 (2363845 -> 2363600), plus 5 flat keeps that bring a
function's instruction multiset to exact parity with retail. MATCH unchanged
at 6931/8235 first-party (84.2%). Over-window unchanged at 2 (inherited,
mdlEffect.c). Build reproduces retail: loadable image
9203646d9aa48ff24eb4ba4b328b02df468a9483, SLUS_216.21
3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb. 22484 call relocations audited,
0 wrong call targets. 92 tests pass.

Keeps

  y_misc      FUN_00435c00    6255 -> 6185  -70  alias return long -> _f32_u32
  br_hero     brHero00263db0  2319 -> 2262  -57  unrolled calls -> loop
  mt_evtMain  FUN_00369a20    6697 -> 6646  -51  extern u32 -> f32
  y_misc      FUN_0042b050     897 ->  865  -32  bare calls -> typed _y2 aliases
  y_misc      FUN_00435440     363 ->  332  -31  same alias, 2 sites
  fclShopMisc FUN_003f1dc0     665 ->  659   -6  callee parameter -> s16
  br_data     func_001fb560    flat, div 0/0 divu 2/2 exact parity
  bpp_panel   FUN_0022f3b0     flat, exact parity
  k_spipe     func_0019ab80    flat, div 0/0 divu 1/1 exact parity
  k_unit      func_001cf940    flat, div 0/0 divu 1/1 exact parity
  datCalendar func_0018c150    flat, exact parity

brHero00263db0's instruction multiset is now EXACT -- the mnemonic census
reports zero differing rows. FUN_0042b050 disappeared from the semantic-ranked
rows entirely; its object shrank 72 bytes.

New tooling

tools/signature_census.py -- classifies all 1267 first-party NONMATCHING
functions by dominant semantic delta, with the remedy attached to each family.
Alignment-free, so meaningful in the badly-aligned functions where most of the
residual sits. 793 functions have a nonzero semantic delta, covering 1,006,289
nd:

  NARROW  499 fn  513,442 nd   XFER  129 fn  235,806 nd
  ARITH    74 fn  148,463 nd   MACFUSE 39 fn  55,050 nd
  CVT      16 fn   22,341 nd   CALL   13 fn   16,724 nd
  DIVMOD   20 fn   11,703 nd   NEG     3 fn    2,760 nd

The finding

Every keep this wave is a declaration or structure fix derived from retail's
instruction stream. Every one of the 25 refusals changed something chosen by
eye. The residual is not mysterious codegen; it is wrong types in the
declaration environment, and the census points at them.

Five variants, all now evidenced:

  mistyped extern global     mt_evtMain fGpffff80d0 declared u32, used as
                             float; every read emitted mtc1 + cvt.s.w
  mis-selected alias         y_misc: a correctly-typed sibling already existed
                             in the file and call sites used the wrong one.
                             Three keeps, -133 total
  mistyped callee parameter  fclShopMisc: retail truncates the argument on the
                             way in because the callee takes s16
  wrong division signedness  five functions, exact div/divu parity
  unrolled calls             br_hero: three explicit pairs where retail loops

Mechanism preconditions, all measured

TRANSFORMATIONS ARE ATOMIC. One unsigned operand promotes a whole expression:
datCalendar needed all four of hp/maxHp/sp/maxSp, and hp/sp alone and
maxHp/maxSp alone each left the census completely unmoved. br_data needed both
its operands. y_misc needed all seven, two and six call sites respectively. A
partial conversion is a mixed spelling matching neither side and measures flat.

A CALLEE-PARAMETER FIX NEEDS >=3 CALL SITES. func_0016f810 at three
occurrences closed three pairs for -6; datGetEquipmentEffect at four
occurrences breached the window; single-occurrence callees breached. The
per-call truncation is a fixed cost that must be amortised against the pairs
closed.

C89 IMPLICIT-INT ORDERING. A call appearing before its alias declaration is
parsed as returning int, and a later declaration then conflicts and fails to
compile. The alias must be declared at file top; only early-scope forms work.

CENSUS-MOVED AND CENSUS-UNMOVED FLATS ARE DIFFERENT RESULTS. The first means
the mechanism works and more sites are needed; the second means the wrong
thing was changed. Three refusals could not be classified because the counts
were not reported.

NARROW is two populations, not one

dsll32 PAIRED with dsra32 is a 64-bit sign-extension -- the narrow-local
defect, fixed by widening a declaration. dsll32 UNPAIRED is a
shift-left-by-32 building a 64-bit value from two halves -- packing, which no
declaration change touches. h_campUpdateSystemMenuTask_148 reads dsll32 -69
with dsra32 only -2: 67 unpaired, entirely packing, and every declaration in it
was already s32 or wider.

NarrowCampW471 classified all 34 Camp NARROW rows on this axis; NarrowYMiscW471
classified all 54 y_misc rows (39 sign-extension, 6 packing, 2 sign-width,
7 shift-only). That splits the 499-function family into two populations with
opposite remedies.

A third sub-case surfaced late: dsll32 + dsrl32 is a ZERO-extension.
btlTarget's alias probe added seven of those while leaving all fourteen
sign-extends untouched, which is why it failed. The census lumps all four
mnemonics into NARROW; split the pairs by partner before choosing a fix.

Characterisations without edits

y_misc FUN_0043ca30, the catalogue's largest residual at 35,545 nd and never
moved in 23 waves: 50 raw IEEE bit-pattern literals passed as float arguments
to FUN_0044a110/FUN_0044a240 (0xc2c80000, 0x43e20000, 0x43880000 and others),
where retail loads the same patterns with lui/mtc1. That is why cvt.s.w is -30.
Correctly left alone under the literal ban -- and it explains 23 waves of
failure on that function.

mdlEffect: DAT_007cae18 and DAT_007cae94 are declared u32, assigned to f32,
and loaded by retail with lwc1. A correctly-typed DAT_007cae18_f32 alias
already exists and one use site does not use it. Using it moved cvt.s.w 7->5
and shrank the object 56 bytes -- and nd went +86. A real, proven, mistyped
global we cannot currently take.

bsa_main bsaMain00210d90: cvt.w.s +82 with mfc1 +82, exactly paired, against
2,944 bytes of deficit. The un-cached-alpha hypothesis was tested and the
census did not move at all, so the conversion pass is elsewhere. Reconstruction
tier declared.

H_Snd_00109180: 260-byte retail tail containing channelIndex*308 dispatch
arithmetic absent from our switch. Reconstruction tier declared.

btlAction btlActionUpdateStateSkill (dsll32 -20/dsra32 -19, nd 11,166),
btlCamera func_002ab330 (9 fused MAC sites) and FUN_002b0280 (6), all with site
counts verified against retail and never attempted -- documented non-attempts
with the analysis already done.

Negative audits that cap mechanisms

Five lanes grepped their files for integer-typed fGp externs: only mt_evtMain
had one. That caps how much of the 129-function XFER family the
mistyped-global mechanism can explain.

Kosaka (14 files) and PanelOp (12 files) audited every FUN alias for
mis-selected siblings: none. The alias defect is concentrated in y_misc.

Process

One orchestrator-caused destructive accident: a batch restore raced an
in-flight lane patch in h_camp.c, invalidating the lane's line mapping
mid-edit and corrupting one line. Reverted. The restore protocol now cuts both
ways -- I will not restore a file while a lane has an unreported edit
outstanding.

One clobbered declaration caught at gate audit: a lane's added prototype
replaced void FUN_003eff00(u64, u8*) in fclShopMisc.c instead of inserting.
Repaired; the prototype is restored and the new line inserted after it.

Artifacts: build/wip/W471_signature.md, W471_brief.md, ten lane reports,
build/baselines/W471_gate.json.

+++ src/Event/mt_evtMain.c
@@ -372,7 +372,7 @@ extern u8 *PTR_s_SPEED_007cd228;
 extern u8 *PTR_s_START_0069e100;
 extern u8 *PTR_s_START_0069ee24;
 extern u8 *PTR_s_START_007ccf50;
-extern u32 fGpffff80d0;
+extern f32 fGpffff80d0;
 extern u32 fGpffff8294;
 extern u32 fGpffff8438;
 extern u8 gp0xffff9d98;
